fixed a bunch of signed/unsigned mismatch warnings in newer gcc versions (mostly related to size_t being unsigned)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@3065 d7cf8633-e32d-0410-b094-e92efae38249

+++ common.c
@@ -339,7 +339,7 @@ char *MSG_ReadString (void)
 {
 	static char string[2048];
 	int l,c;
-	for (l = 0;l < sizeof(string) - 1 && (c = MSG_ReadChar()) != -1 && c != 0;l++)
+	for (l = 0;l < (int) sizeof(string) - 1 && (c = MSG_ReadChar()) != -1 && c != 0;l++)
 		string[l] = c;
 	string[l] = 0;
 	return string;
